% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datacamp_slides.R
\name{download_datacamp_slides}
\alias{download_datacamp_slides}
\title{Easily download DataCamp slides}
\usage{

  download_datacamp_slides(input = "https://s3.amazonaws.com/assets.datacamp.com/production/course_15268/slides/chapter{1:4}.pdf",
  dir = "refs")
}
\arguments{
\item{input}{character}

\item{dir}{character}
}
\description{
Easily download DataCamp slides
}
